+++ crates/nostr-relay-pool/src/relay/error.rs
@@ -61,6 +61,11 @@ pub enum Error {
     NotConnected,
     /// Relay banned
     Banned,
+    /// Connection rejected
+    ConnectionRejected {
+        /// Reason
+        reason: Option<String>,
+    },
     /// Received termination request
     TerminationRequest,
     /// Received shutdown
@@ -143,6 +148,10 @@ impl fmt::Display for Error {
             Self::NotReady => write!(f, "relay is initialized but not ready"),
             Self::NotConnected => write!(f, "relay not connected"),
             Self::Banned => write!(f, "relay banned"),
+            Self::ConnectionRejected { reason } => {
+                let reason: &str = reason.as_deref().unwrap_or("unknown");
+                write!(f, "connection rejected: reason={reason}")
+            }
             Self::TerminationRequest => write!(f, "received termination request"),
             Self::ReceivedShutdown => write!(f, "received shutdown"),
             Self::RelayMessage(message) => write!(f, "{message}"),